[TMGimporter]Minor update from August 2018[v51]
* Merged files

* almost working tmg glade selection gui

*** non functional***
