+++ TMGimporter/README.md
@@ -1,0 +1,40 @@
+Work in progress - unreleased addon.
+
+# TMG Importer #
+
+TMG Importer is a Gramps genealogy program addon to Import from the Project backup file
+(*.SQZ) created by the last version (v9.05.000 released:3 Dec 2014) of "The Master Genealogist (TMG)" genealogy program from Whollygenes discontinued in 2014.
+
+This addon only supports importing from an TMG Backup (*.SQZ) and not directly from the file system eg: PJC / TMG / VER.
+
+### How do I make it work? ###
+
+* Requires you install the required prerequiste "dbf ( https://pypi.python.org/pypi/dbf )" python package by Ethan Furman (version 0.96.8 or greater). This reads TMG's Database(*.DBF) files
+
+* Install the "TMG Importer" using the Gramps addon manager.
+
+* Start Gramps and create a new Family tree.
+
+* On the menu select Family Trees > Import... from the menu and then find the location of your TMG Backup file (*.SQZ) select the Import button.
+
+* If you have more than one Dataset in the project, select the one you want to import.
+
+### Have an issue? ###
+
+Please raise any issue on the Gramps Bug tracker.
+
+https://gramps-project.org/wiki/index.php?title=Addon:TMGimporter
+
+### Alternative Import Options. ###
+
+Import from another genealogy program - The Master Genealogist (TMG)
+https://gramps-project.org/wiki/index.php/Import_from_another_genealogy_program#The_Master_Genealogist_.28TMG.29
+
+### External links ###
+
+http://www.whollygenes.com/
+
+http://www.whollygenes.com/forums201/index.php?/topic/15798-tmg-to-be-discontinued/
+
+https://en.wikipedia.org/wiki/The_Master_Genealogist
+
