+++ TMGimporter/importtmg.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 #
-# Gramps - a GTK+/GNOME based genealogy program
+# TMG Importer addon for Gramps genealogy program
 #
-# Copyright (C) 2017 Sam Manzi
+# Copyright (C) 2017-2018 Sam Manzi
 #
 # This program is free software; you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -27,18 +27,20 @@ Import from an Whollygenes - The Master Genealogist (TMG) Project backup file
 
 #-------------------------------------------------------------------------
 #
-# Standard python modules
+# Standard Python modules
 #
 #-------------------------------------------------------------------------
-
-
-#------------------------------------------------------------------------
-#
-# Set up logging
-#
-#------------------------------------------------------------------------
 import logging
 LOG = logging.getLogger(".TMGImport")
+
+#-------------------------------------------------------------------------
+#
+# Gramps modules
+#
+#-------------------------------------------------------------------------
+from gramps.gen.const import GRAMPS_LOCALE as glocale
+_ = glocale.translation.gettext
+from gramps.gen.const import URL_WIKISTRING
 
 #-------------------------------------------------------------------------
 #
@@ -46,7 +48,7 @@ LOG = logging.getLogger(".TMGImport")
 #
 #-------------------------------------------------------------------------
 
-from lib import libtmg
+import libtmg
 
 #-------------------------------------------------------------------------
 #
@@ -58,7 +60,82 @@ def importSqzData(database, filename, user):
     Function called by Gramps to import TMG project dataset(s).
     Uses TMG library to process .SQZ Backup file
     """
+    print("-----------------------------------------------------")
+    print("# TMG Importer : Started : Processing (*.SQZ) file  #")
+    print("-----------------------------------------------------")
+    #tmgabortimport = False   # Import Status
     importer = libtmg.importData
     info = importer(database, filename, user)
-
+    #print("tmgabortimport : ", tmgabortimport)
+    print("-----------------------------------------------------")
+    print("# TMG Importer : Finished : Processing (*.SQZ) file #")
+    print("-----------------------------------------------------")
     return info
+
+#-------------------------------------------------------------------------
+#
+# Unsupported TMG formats. Imported data must be in a TMG *.SQZ backup file
+#
+#-------------------------------------------------------------------------
+def importpjc(database, filename, user):
+    '''
+    *.PJC - Project Configuration File for TMG 5.0 to TMG 9.05
+    '''
+    import_dict = { 'gramps_wiki_import_pjc_direct_url' :
+                         URL_WIKISTRING +
+                             "Addon:TMGimporter#"
+                             "Unsupported_formats" }
+    user.notify_error(_("%s could not be opened") % filename,
+                      _("Directly importing from TMG files is not supported\n"
+                        "by the TMG Importer Addon.\n\n"
+                        "You need to use an backup copy of the TMG project(*.sqz)\n\n"
+                        "Ensure that your TMG project was created by:\n"
+                        "TMG version 5.x or greater.\n\n"
+                        "Your file:\n"
+                        "*.PJC - Project Configuration File for TMG 5.0 to TMG 9.05\n\n"
+                        "Please refer to:\n"
+                        "%(gramps_wiki_import_pjc_direct_url)s" ) %
+                                import_dict )
+    return
+
+def importtmg(database, filename, user):
+    '''
+    *.TMG - Version Control File for TMG 2.0 to TMG 4.0d
+    '''
+    import_dict = { 'gramps_wiki_import_pjc_direct_url' :
+                         URL_WIKISTRING +
+                             "Addon:TMGimporter#"
+                             "Unsupported_formats" }
+    user.notify_error(_("%s could not be opened") % filename,
+                      _("Directly importing from TMG files is not supported"
+                        "by the TMG Importer Addon."
+                        "You need to use an backup copy of TMG project(*.sqz)"
+                        "Ensure that your TMG project was created by:"
+                        "TMG version 5.x or greater."
+                        "Your file:"
+                        "*.TMG - Version Control File for TMG 2.0 to TMG 4.0d"
+                        "Please refer to:"
+                        "%(gramps_wiki_import_pjc_direct_url)s" ) %
+                                import_dict )
+    return
+
+def importver(database, filename, user):
+    '''
+    *.VER - Version Control File for TMG 1.2 and earlier
+    '''
+    import_dict = { 'gramps_wiki_import_pjc_direct_url' :
+                         URL_WIKISTRING +
+                             "Addon:TMGimporter#"
+                             "Unsupported_formats" }
+    user.notify_error(_("%s could not be opened") % filename,
+                      _("Directly importing from TMG files is not supported"
+                        "by the TMG Importer Addon."
+                        "You need to use an backup copy of TMG project(*.sqz)"
+                        "Ensure that your TMG project was created by:"
+                        "TMG version 5.x or greater."
+                        "Your file:"
+                        "*.VER - Version Control File for TMG 1.2 and earlier"
+                        "Please refer to:"
+                        "%(gramps_wiki_import_pjc_direct_url)s" ) %
+                                import_dict )
+    return
