--- TMGimporter/po/de-local.po
@@ -1,114 +1,0 @@
-# German translation for Gramps
-# This file is distributed under the same license as the Gramps package.
-# translation of de.po to Deutsch
-#
-#
-# Anton Huber <anton_huber@gmx.de>, 2005,2006.
-# Sebastian Vöcking <sebastian_voecking@gmx.de>, 2005.
-# Sebastian Vöcking <voeck@web.de>, 2005.
-# Martin Hawlisch <martin@hawlisch.de>, 2005, 2006.
-# Alex Roitman <shura@gramps-project.org>, 2006.
-# Mirko Leonhäuser <mirko@leonhaeuser.de>, 2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018.
-# Alois Pöttker <alois.poettker@gmx.de>, 2017.
-msgid ""
-msgstr ""
-"Project-Id-Version: de\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2018-05-03 15:48+0200\n"
-"PO-Revision-Date: 2018-05-04 00:24+0100\n"
-"Last-Translator: Mirko Leonhäuser <mirko@leonhaeuser.de>\n"
-"Language-Team: German <kde-i18n-de@kde.org>\n"
-"Language: de\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"X-Generator: Lokalize 2.0\n"
-"Plural-Forms: nplurals=2; plural=(n != 1);\n"
-
-#: TMGimporter/importtmg.gpr.py:71
-msgid "TMG Project Backup"
-msgstr "TMG Projekt Sicherung"
-
-#: TMGimporter/importtmg.gpr.py:72 TMGimporter/importtmg.gpr.py:85
-#: TMGimporter/importtmg.gpr.py:98 TMGimporter/importtmg.gpr.py:111
-msgid "Import TMG project files"
-msgstr "TMG Projektdateien importieren"
-
-#: TMGimporter/importtmg.gpr.py:84 TMGimporter/importtmg.gpr.py:97
-#: TMGimporter/importtmg.gpr.py:110
-msgid "TMG Unsupported"
-msgstr "TMG nicht unterstützt"
-
-#: TMGimporter/importtmgunsupported.py:55
-#: TMGimporter/importtmgunsupported.py:76
-#: TMGimporter/importtmgunsupported.py:97
-#, python-format
-msgid "%s could not be opened"
-msgstr "%s konnte nicht geöffnet werden"
-
-#: TMGimporter/importtmgunsupported.py:56
-#, python-format
-msgid ""
-"Directly importing from TMG files is not supported\n"
-"by the TMG Importer Addon.\n"
-"\n"
-"You need to use an backup copy of the TMG project(*.sqz)\n"
-"\n"
-"Ensure that your TMG project was created by:\n"
-"TMG version 5.x or greater.\n"
-"\n"
-"Your file:\n"
-"*.PJC - Project Configuration File for TMG 5.0 to TMG 9.05\n"
-"\n"
-"Please refer to:\n"
-"%(gramps_wiki_import_pjc_direct_url)s"
-msgstr ""
-"Direkter Import von TMG Dateien wird nicht unterstützt\n"
-"durch die TMG Import Erweiterung.\n"
-"\n"
-"Du benötigst eine Sicherungskopie des TMG Projekt(*.sqz)\n"
-"\n"
-"Stelle sicher das dein TMG Projekt erstellt wurde mit:\n"
-"TMG Version 5.x oder höher.\n"
-"\n"
-"Deine Datei:\n"
-"*.PJC - Projekt Konfiguration Datei für TMG 5.0 nach TMG 9.05\n"
-"\n"
-"Bitte beachte:\n"
-"%(gramps_wiki_import_pjc_direct_url)s"
-
-#: TMGimporter/importtmgunsupported.py:77
-#, python-format
-msgid ""
-"Directly importing from TMG files is not supportedby the TMG Importer Addon."
-"You need to use an backup copy of TMG project(*.sqz)Ensure that your TMG "
-"project was created by:TMG version 5.x or greater.Your file:*.TMG - Version "
-"Control File for TMG 2.0 to TMG 4.0dPlease refer to:"
-"%(gramps_wiki_import_pjc_direct_url)s"
-msgstr ""
-"Der direkte Import von TMG-Dateien wird vom TMG Importer Addon nicht"
-" unterstützt. Du musst eine Sicherungskopie des TMG-Projekts (* .sqz)"
-" verwenden. Stelle sicher, dass dein TMG"
-"Projekt erstellt wurde von: TMG Version 5.x oder höher. Deine Datei: *."
-" TMG-Version"
-"Steuerdatei für TMG 2.0 bis TMG 4.0d Bitte beachte:"
-"%(gramps_wiki_import_pjc_direct_url)s"
-
-#: TMGimporter/importtmgunsupported.py:98
-#, python-format
-msgid ""
-"Directly importing from TMG files is not supportedby the TMG Importer Addon."
-"You need to use an backup copy of TMG project(*.sqz)Ensure that your TMG "
-"project was created by:TMG version 5.x or greater.Your file:*.VER - Version "
-"Control File for TMG 1.2 and earlierPlease refer to:"
-"%(gramps_wiki_import_pjc_direct_url)s"
-msgstr ""
-"Der direkte Import von TMG-Dateien wird vom TMG Importer Addon nicht"
-" unterstützt. "
-"Du musst eine Sicherungskopie des TMG-Projekts (* .sqz) verwenden. Stelle"
-" sicher, dass dein TMG"
-"Projekt erstellt wurde von: TMG Version 5.x oder höher. Deine Datei: *."
-" VER-Version"
-"Steuerdatei für TMG 1.2 und früher. Bitte beachte:"
-"%(gramps_wiki_import_pjc_direct_url)s"
-
