--- TMGimporter/po/hr-local.po
@@ -1,102 +1,0 @@
-# Croatian translation for Gramps
-# This file is distributed under the same license as the Gramps package.
-# Copyright (C) Gramps project
-# Milo Ivir <mail@milotype.de>, 2019.
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Gramps 5.x\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2019-01-04 18:55-0600\n"
-"PO-Revision-Date: 2019-03-08 11:56+0100\n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"X-Generator: Poedit 2.2.1\n"
-"Last-Translator: Milo Ivir <mail@milotype.de>\n"
-"Plural-Forms: nplurals=3; plural=(n%10==1 && n%100!=11 ? 0 : n%10>=2 && n"
-"%10<=4 && (n%100<12 || n%100>14) ? 1 : 2);\n"
-"Language: hr\n"
-
-#: TMGimporter/importtmg.gpr.py:72
-msgid "TMG Project Backup"
-msgstr "Sigurnosna kopija TMG projekta"
-
-#: TMGimporter/importtmg.gpr.py:73 TMGimporter/importtmg.gpr.py:86
-#: TMGimporter/importtmg.gpr.py:99 TMGimporter/importtmg.gpr.py:112
-msgid "Import TMG project files"
-msgstr "Uvezi datoteke TMG projekta"
-
-#: TMGimporter/importtmg.gpr.py:85 TMGimporter/importtmg.gpr.py:98
-#: TMGimporter/importtmg.gpr.py:111
-msgid "TMG Unsupported"
-msgstr "TMG nije podržan"
-
-#: TMGimporter/importtmgunsupported.py:55
-#: TMGimporter/importtmgunsupported.py:76
-#: TMGimporter/importtmgunsupported.py:97
-#, python-format
-msgid "%s could not be opened"
-msgstr "Nije bilo moguće otvoriti „%s”"
-
-#: TMGimporter/importtmgunsupported.py:56
-#, python-format
-msgid ""
-"Directly importing from TMG files is not supported\n"
-"by the TMG Importer Addon.\n"
-"\n"
-"You need to use an backup copy of the TMG project(*.sqz)\n"
-"\n"
-"Ensure that your TMG project was created by:\n"
-"TMG version 5.x or greater.\n"
-"\n"
-"Your file:\n"
-"*.PJC - Project Configuration File for TMG 5.0 to TMG 9.05\n"
-"\n"
-"Please refer to:\n"
-"%(gramps_wiki_import_pjc_direct_url)s"
-msgstr ""
-"Dodatak za uvoz TMG-a ne podržava \n"
-"direktan uvoz TMG-datoteka. \n"
-"\n"
-"Moraš koristiti sigurnosnu kopiju TMG projekta (* .sqz). \n"
-"\n"
-"Provjeri, da je tvoj TMG projekt stvoren s: \n"
-"TMG inačicom 5.x ili višom. \n"
-"\n"
-"Tvoja datoteka: \n"
-"*.PJC - konfiguracijska datoteka projekta za TMG 5.0 to TMG 9.05 \n"
-"\n"
-"Obrati se na: \n"
-"%(gramps_wiki_import_pjc_direct_url)s"
-
-#: TMGimporter/importtmgunsupported.py:77
-#, python-format
-msgid ""
-"Directly importing from TMG files is not supportedby the TMG Importer Addon."
-"You need to use an backup copy of TMG project(*.sqz)Ensure that your TMG "
-"project was created by:TMG version 5.x or greater.Your file:*.TMG - Version "
-"Control File for TMG 2.0 to TMG 4.0dPlease refer to:"
-"%(gramps_wiki_import_pjc_direct_url)s"
-msgstr ""
-"Dodatak za uvoz TMG-a ne podržava direktan uvoz TMG-datoteka. Za uvoz moraš "
-"koristiti sigurnosnu kopiju TMG projekta (* .sqz). Provjeri, da je tvoj TMG "
-"projekt stvoren s: TMG inačicom 5.x ili višom. Tvoja datoteka: *.TMG - "
-"upravljačka datoteka inačica za TMG 2.0 to TMG 4.0d. Obrati se na: "
-"%(gramps_wiki_import_pjc_direct_url)s"
-
-#: TMGimporter/importtmgunsupported.py:98
-#, python-format
-msgid ""
-"Directly importing from TMG files is not supportedby the TMG Importer Addon."
-"You need to use an backup copy of TMG project(*.sqz)Ensure that your TMG "
-"project was created by:TMG version 5.x or greater.Your file:*.VER - Version "
-"Control File for TMG 1.2 and earlierPlease refer to:"
-"%(gramps_wiki_import_pjc_direct_url)s"
-msgstr ""
-"Dodatak za uvoz TMG-a ne podržava direktan uvoz TMG-datoteka. Za uvoz moraš "
-"koristiti sigurnosnu kopiju TMG projekta (* .sqz). Provjeri, da je tvoj TMG "
-"projekt stvoren s: TMG inačicom 5.x ili višom. Tvoja datoteka: *.VER - "
-"upravljačka datoteka inačica za TMG 1.2 i prijašnje. Obrati se na: "
-"%(gramps_wiki_import_pjc_direct_url)s"
