--- TMGimporter/po/template.pot
@@ -1,77 +1,0 @@
-# SOME DESCRIPTIVE TITLE.
-# Copyright (C) YEAR THE PACKAGE'S COPYRIGHT HOLDER
-# This file is distributed under the same license as the PACKAGE package.
-# FIRST AUTHOR <EMAIL@ADDRESS>, YEAR.
-#
-#, fuzzy
-msgid ""
-msgstr ""
-"Project-Id-Version: PACKAGE VERSION\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2019-09-19 10:57-0500\n"
-"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
-"Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
-"Language-Team: LANGUAGE <LL@li.org>\n"
-"Language: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=CHARSET\n"
-"Content-Transfer-Encoding: 8bit\n"
-
-#: TMGimporter/importtmg.gpr.py:71
-msgid "TMG Project Backup"
-msgstr ""
-
-#: TMGimporter/importtmg.gpr.py:72 TMGimporter/importtmg.gpr.py:85
-#: TMGimporter/importtmg.gpr.py:98 TMGimporter/importtmg.gpr.py:111
-msgid "Import TMG project files"
-msgstr ""
-
-#: TMGimporter/importtmg.gpr.py:84 TMGimporter/importtmg.gpr.py:97
-#: TMGimporter/importtmg.gpr.py:110
-msgid "TMG Unsupported"
-msgstr ""
-
-#: TMGimporter/importtmgunsupported.py:55
-#: TMGimporter/importtmgunsupported.py:76
-#: TMGimporter/importtmgunsupported.py:97
-#, python-format
-msgid "%s could not be opened"
-msgstr ""
-
-#: TMGimporter/importtmgunsupported.py:56
-#, python-format
-msgid ""
-"Directly importing from TMG files is not supported\n"
-"by the TMG Importer Addon.\n"
-"\n"
-"You need to use an backup copy of the TMG project(*.sqz)\n"
-"\n"
-"Ensure that your TMG project was created by:\n"
-"TMG version 5.x or greater.\n"
-"\n"
-"Your file:\n"
-"*.PJC - Project Configuration File for TMG 5.0 to TMG 9.05\n"
-"\n"
-"Please refer to:\n"
-"%(gramps_wiki_import_pjc_direct_url)s"
-msgstr ""
-
-#: TMGimporter/importtmgunsupported.py:77
-#, python-format
-msgid ""
-"Directly importing from TMG files is not supportedby the TMG Importer Addon."
-"You need to use an backup copy of TMG project(*.sqz)Ensure that your TMG "
-"project was created by:TMG version 5.x or greater.Your file:*.TMG - Version "
-"Control File for TMG 2.0 to TMG 4.0dPlease refer to:"
-"%(gramps_wiki_import_pjc_direct_url)s"
-msgstr ""
-
-#: TMGimporter/importtmgunsupported.py:98
-#, python-format
-msgid ""
-"Directly importing from TMG files is not supportedby the TMG Importer Addon."
-"You need to use an backup copy of TMG project(*.sqz)Ensure that your TMG "
-"project was created by:TMG version 5.x or greater.Your file:*.VER - Version "
-"Control File for TMG 1.2 and earlierPlease refer to:"
-"%(gramps_wiki_import_pjc_direct_url)s"
-msgstr ""
